apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.2
websockets==13.1
